# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Pantry API",
    "description": "The REST API of pantry allows us to manage Users and their accounts, orders, favourite salads and custom salads, Sally owner accounts,  Salads, Ingredients, User Ratings",
    "termsOfService": "http://casabots.com/terms/",
    "license": {
      "name": "MIT",
      "url": "http://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8080",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/text"
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Retrieve a list of users",
        "operationId": "getUsers",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "users response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "default": {
            "description": "unexpected error"
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Creates a new user",
        "operationId": "createUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful user creation"
          },
          "default": {
            "description": "unexpected error"
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Gets a user based on user id",
        "operationId": "getUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "the id of user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "user response",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "default": {
            "description": "unexpected error"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "Removes a user based on user id",
        "operationId": "removeUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "the id of user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "user deletion success"
          },
          "default": {
            "description": "unexpected error"
          }
        }
      }
    },
    "/users/update": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Updates a user",
        "operationId": "updateUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully updated user"
          },
          "default": {
            "description": "unexpected error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "dob": {
          "type": "string",
          "format": "date"
        },
        "location": {
          "$ref": "#/definitions/GeoLocation"
        },
        "savedSalads": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CustomSalad"
          }
        },
        "favouriteSalads": {
          "$ref": "#/definitions/SaladList"
        }
      }
    },
...
    "CustomSalad": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "ingredients": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ingredient"
          }
        }
      }
    },
    "Ingredient": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "quantity": {
          "type": "string"
        }
      }
    },
...
    "SaladList": {
      "type": "object",
      "properties": {
        "predefinedSalads": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "customSalads": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CustomSalad"
          }
        }
      }
    },
...
    "GeoLocation": {
      "type": "object",
      "properties": {
        "latitude": {
          "type": "number",
          "format": "double"
        },
        "longitude": {
          "type": "number",
          "format": "double"
        }
      }
    }
  }
}